server/requirements.txt

flask>=2.0.0
flask-cors>=3.0.0
openai>=1.0.0
anthropic>=0.8.0
google-generativeai>=0.3.0
python-dotenv>=0.19.0
